Update eslint (major v1.0.0, new methods, yummy!). An initial exploration of the actual implamentation. Added test with several levels of intricacy (testing normal vars vs class props)
