+++ index.js
@@ -1,6 +1,7 @@
 module.exports = {
   rules: {
     'space': require('./lib/rules/space'),
-    'name': require('./lib/rules/name')
+    'name': require('./lib/rules/name'),
+    'properties': require('./lib/rules/properties')
   }
 };
